// .keystone/schema-types.ts
type Scalars = {
  readonly ID: string;
  readonly Boolean: boolean;
  readonly String: string;
  readonly Int: number;
  readonly Float: number;
  readonly JSON: import('@keystone-next/types').JSONValue;
};
// ...
export type UserWhereUniqueInput = {
  readonly id: Scalars['ID'];
};
// ...
export type UserCreateInput = {
  readonly name?: Scalars['String'] | null;
  readonly email?: Scalars['String'] | null;
  readonly password?: Scalars['String'] | null;
  readonly cart?: CartItemRelateToManyInput | null;
  readonly orders?: OrderRelateToManyInput | null;
  readonly role?: RoleRelateToOneInput | null;
  readonly products?: ProductRelateToManyInput | null;
  readonly passwordResetToken?: Scalars['String'] | null;
  readonly passwordResetIssuedAt?: Scalars['String'] | null;
  readonly passwordResetRedeemedAt?: Scalars['String'] | null;
  readonly magicAuthToken?: Scalars['String'] | null;
  readonly magicAuthIssuedAt?: Scalars['String'] | null;
  readonly magicAuthRedeemedAt?: Scalars['String'] | null;
};

export type UsersCreateInput = {
  readonly data?: UserCreateInput | null;
};

export type UserRelateToOneInput = {
  readonly create?: UserCreateInput | null;
  readonly connect?: UserWhereUniqueInput | null;
  readonly disconnect?: UserWhereUniqueInput | null;
  readonly disconnectAll?: Scalars['Boolean'] | null;
};
// ...
export type OrderItemWhereUniqueInput = {
  readonly id: Scalars['ID'];
};
// ...
export type OrderItemCreateInput = {
  readonly name?: Scalars['String'] | null;
  readonly description?: Scalars['String'] | null;
  readonly price?: Scalars['Int'] | null;
  readonly photo?: ProductImageRelateToOneInput | null;
  readonly quantity?: Scalars['Int'] | null;
  readonly order?: OrderRelateToOneInput | null;
};

export type OrderItemsCreateInput = {
  readonly data?: OrderItemCreateInput | null;
};

export type OrderItemRelateToManyInput = {
  readonly create?: ReadonlyArray<OrderItemCreateInput | null> | null;
  readonly connect?: ReadonlyArray<OrderItemWhereUniqueInput | null> | null;
  readonly disconnect?: ReadonlyArray<OrderItemWhereUniqueInput | null> | null;
  readonly disconnectAll?: Scalars['Boolean'] | null;
};
// ...
export type OrderCreateInput = {
  readonly total?: Scalars['Int'] | null;
  readonly items?: OrderItemRelateToManyInput | null;
  readonly user?: UserRelateToOneInput | null;
  readonly charge?: Scalars['String'] | null;
};